# allow tensordot to handle 32+ 'involved' dimensions