This enables ``da.tensordot`` to handle contractions with more than 32 'involved' dimensions - i.e. when the number of unique left dimensions + unique right dimensions + contracted dimensions > 32, which is a much more stringent condition than simply the maximum dimension of any array (#5595).

The fix seems quite simple so I do wonder whether there was some reasoning behind not letting ``blockwise`` perform the sum itself, but possibly the particular explicit sum I'm removing just predates blockwise (#2186).

- [x] Tests added / passed
- [x] Passes `black dask` / `flake8 dask`