@TomAugspurger 2020-07-06
@jakirkham do you have any thoughts here?