@mrocklin 2020-07-06
My guess is that this will hurt performance.  It forces Dask to bring many inputs together at once in order to perform a tensordot on all of them.  This may not fit well in memory.

The implementation in master calls tensordot between pairs of chunks, then reduces those chunks with a sum, and then aggregates them with a tree reduction (all in the sum all). The implementation here brings all of the data necessary to create an output-chunk together in one machine and then does the full tensordot computation.  This isn't always possible.