@TomAugspurger 2020-07-06
@jcmgray are you able to construct a benchmark to see if this hurts performance / memory usage for large arrays with many blocks?