@jcmgray 2020-07-06
> The implementation in master calls tensordot between pairs of chunks, then reduces those chunks with a sum, and then aggregates them with a tree reduction (all in the sum all). The implementation here brings all of the data necessary to create an output-chunk together in one machine and then does the full tensordot computation. This isn't always possible.

To be sure I am not really certain of what ``blockwise`` does - I had just copied what ``matmul`` does (though of course this only contracts 1 index) and hoped that what was happening was that the sum reduction on axes was turned into an actual sum over chunks without forming the outer product array first.

I set up this benchmark:

```python
import numpy as np
import dask.array as da

np.random.seed(42)

d = 2000       # dimension size
n = 2          # number of dimensions on inputs
nmax = 2       # maximum number of output dimensions
c = 100        # chunk size
dchunked = 1   # number of dimensions to chunk

nx = np.random.normal(size=[d]* n) + 1j * np.random.normal(size=[d]* n)
ny = np.random.normal(size=[d]* n) + 1j * np.random.normal(size=[d]* n)


xchunks = [-1] * n
ychunks = [-1] * n

for d in np.random.choice(np.arange(n), size=dchunked, replace=False):
    xchunks[d] = c
for d in np.random.choice(np.arange(n), size=dchunked, replace=False):
    ychunks[d] = c

dx = da.from_array(nx, chunks=xchunks)
dy = da.from_array(ny, chunks=ychunks)
```
and timed this:
```python
%%time

import warnings
import tqdm

warnings.filterwarnings("ignore")

for _ in tqdm.trange(100):
    # number of dimensions to contract
    nax = np.random.randint(
        max(0, (2 * n - nmax) // 2), 
        n + 1
    )

    # which dimensions to contract
    lax = tuple(np.random.choice(np.arange(n), size=nax, replace=False))
    rax = tuple(np.random.choice(np.arange(n), size=nax, replace=False))
    
    dxy = da.tensordot(dx, dy, [lax, rax]).compute()
```

## Results

These were all run with a ``LocalCluster``  using 4 processes.

```
d = 2          # dimension size
n = 10         # number of dimensions on inputs
nmax = 24      # maximum number of output dimensions
c = 1          # chunk size
dchunked = 5   # number of dimensions to chunk
```
* Previous: Wall time: 1min 17s
* Now: Wall time: 48.2 s

---

```
d = 16         # dimension size
n = 4          # number of dimensions on inputs
nmax = 6       # maximum number of output dimensions
c = 4          # chunk size
dchunked = 2   # number of dimensions to chunk
```
* Previous: Wall time: 1min 19s
* Now: Wall time: 41.5 s

---

```
d = 2000       # dimension size
n = 2          # number of dimensions on inputs
nmax = 2       # maximum number of output dimensions
c = 100        # chunk size
dchunked = 1   # number of dimensions to chunk
```

* Previous: Wall time: 5min 35s (762MB)
* Now: Wall time: Wall time: 2min 11s (667MB)

So compute performance seems better if anything (probably due to the some of the summations being performed within ``np.tensordot`` rather than all in the axes reduction?) 

The memory values on the last set of results I just read-off from the scheduler status page - are these an accurate way to gauge memory usage?