@jcmgray 2021-01-11
Closing this for now, as I don't fully understand the memory implications etc for using ``blockwise``.